+++ Sources/Sentry/SentryEvent.m
@@ -5,6 +5,7 @@
 #import "SentryClient.h"
 #import "SentryDebugMeta.h"
 #import "SentryException.h"
+#import "SentryId.h"
 #import "SentryMeta.h"
 #import "SentryStacktrace.h"
 #import "SentryThread.h"
@@ -23,9 +24,7 @@ NS_ASSUME_NONNULL_BEGIN
 {
     self = [super init];
     if (self) {
-        self.eventId =
-            [[[NSUUID UUID].UUIDString stringByReplacingOccurrencesOfString:@"-"
-                                                                 withString:@""] lowercaseString];
+        self.eventId = [[SentryId alloc] init];
         self.level = level;
         self.platform = @"cocoa";
     }
@@ -48,7 +47,7 @@ NS_ASSUME_NONNULL_BEGIN
     }
 
     NSMutableDictionary *serializedData = @{
-        @"event_id" : self.eventId,
+        @"event_id" : self.eventId.sentryIdString,
         @"timestamp" : [self.timestamp sentry_toIso8601String],
         @"platform" : @"cocoa",
     }