+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -33,6 +33,7 @@
 #import "SentryHttpDateParser.h"
 #import "SentryHttpTransport.h"
 #import "SentryHub+TestInit.h"
+#import "SentryId.h"
 #import "SentryMeta.h"
 #import "SentryRateLimitCategory.h"
 #import "SentryRateLimitCategoryMapper.h"